+++ R/monitor_isTidy.R
@@ -15,9 +15,10 @@
 #' library(AirMonitor)
 #'
 #' mts_monitor <-
-#'   NW_megafires %>%
-#'   monitor_select(c('450d08fb5a3e4ea0_530470009', '40ffdacb421a5ee6_530470010')
-#' )
+#'   NW_Megafires %>%
+#'   AirMonitor::monitor_select(
+#'     c('450d08fb5a3e4ea0_530470009', '40ffdacb421a5ee6_530470010')
+#'   )
 #'
 #' mts_tidy <- monitor_toTidy(mts_monitor)
 #' monitor_isTidy(mts_tidy)