+++ localOldFunctions/theme_dailyBarplot_pwfsl.R
@@ -14,7 +14,7 @@
 #' @import ggplot2
 #' @export
 
-theme_dailyBarplot_pwfsl <- function(base_size = 11,
+theme_dailyBarplot_airfire <- function(base_size = 11,
                                      base_family = "") {
   
   theme_classic(