+++ local_R/PWFSLSmoke.R
@@ -1,7 +1,0 @@
-#' Depend on AirMonitor
-#'
-#' @name AirMonitor
-#' @keywords internal
-#' @import AirMonitor
-NULL
-